# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "ProjectsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns an array of projects."
          }
        },
        "summary": "Get all projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "operationId": "ProjectsController_create",
        "parameters": [
          {
            "name": "x-user-id",
            "required": true,
            "in": "header",
            "description": "ID of the user creating the project",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The project has been successfully created."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "summary": "Create a new project (Requires Project Owner or Administrator)",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single project."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "summary": "Get a project by ID",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The project has been successfully updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "summary": "Update a project (Requires Project Owner or Administrator)",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project has been successfully deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Project not found."
          }
        },
        "summary": "Delete a project (Requires Project Owner or Administrator)",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/owner/{ownerId}": {
      "get": {
        "operationId": "ProjectsController_findByOwnerId",
        "parameters": [
          {
            "name": "ownerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of projects belonging to the owner."
          }
        },
        "summary": "Get projects by Owner ID",
        "tags": [
          "Projects"
        ]
      }
    },
    "/tasks/project/{projectId}": {
      "get": {
        "operationId": "TasksController_findAllByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of tasks."
          }
        },
        "summary": "Get all tasks for a specific project",
        "tags": [
          "Tasks"
        ]
      }
    },
...
    "/communication/messages/project/{projectId}": {
      "get": {
        "operationId": "CommunicationController_getMessagesByProjectId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an array of messages sorted chronologically."
          }
        },
        "summary": "Get chat history for a specific project",
        "tags": [
          "Communication"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter notifications by user ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all notifications or filter by user",
        "tags": [
          "Notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "summary": "Create a new notification",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_findOne",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "summary": "Get a notification by ID",
        "tags": [
          "Notifications"
        ]
      },
      "put": {
        "operationId": "NotificationsController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "summary": "Update a notification",
        "tags": [
          "Notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted successfully"
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "summary": "Delete a notification",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "put": {
        "operationId": "NotificationsController_markAsRead",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Notification ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            }
          },
          "404": {
            "description": "Notification not found"
          }
        },
        "summary": "Mark a notification as read",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "put": {
        "operationId": "NotificationsController_markAllAsRead",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for authorization",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "user"
              ]
            }
          },
          {
            "name": "x-user-id",
            "required": false,
            "in": "header",
            "description": "Optional user ID to mark only that user's notifications as read",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications marked as read",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        },
        "summary": "Mark all notifications as read, optionally filtered by user",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "TeamForge API",
    "description": "The Student Project Collaboration Platform REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "AI Study Planner"
          },
          "description": {
            "type": "string",
            "example": "An intelligent study scheduling app."
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ],
            "example": "Medium"
          },
          "requiredSkills": {
            "example": [
              "React",
              "Python",
              "ML"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "duration": {
            "type": "string",
            "example": "3 Months"
          },
          "collaborators": {
            "example": [
              {
                "id": "1",
                "name": "User"
              }
            ],
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "title",
          "description",
          "difficulty",
          "requiredSkills",
          "duration"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "AI Study Planner"
          },
          "description": {
            "type": "string",
            "example": "An intelligent study scheduling app."
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "Easy",
              "Medium",
              "Hard"
            ],
            "example": "Medium"
          },
          "requiredSkills": {
            "example": [
              "React",
              "Python",
              "ML"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "duration": {
            "type": "string",
            "example": "3 Months"
          },
          "collaborators": {
            "example": [
              {
                "id": "1",
                "name": "User"
              }
            ],
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the notification",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns this notification",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "type": {
            "type": "string",
            "description": "Type of notification",
            "example": "PROJECT_INVITATION",
            "enum": [
              "PROJECT_INVITATION",
              "APPLICATION_APPROVED",
              "TASK_APPROVED",
              "MENTOR_REQUEST",
              "NEW_MESSAGE",
              "PROJECT_UPDATE"
            ]
          },
          "message": {
            "type": "string",
            "description": "Notification message content",
            "example": "You have been invited to join the AI Study Planner project"
          },
          "readStatus": {
            "type": "boolean",
            "description": "Whether the notification has been read",
            "example": false,
            "default": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the notification was created",
            "example": "2024-01-15T10:30:00Z"
          },
          "referenceId": {
            "type": "string",
            "description": "Reference ID (could be project ID, task ID, etc.)",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "message",
          "readStatus",
          "createdAt"
        ]
      },
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID of the user who will receive this notification",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "type": {
            "type": "string",
            "description": "Type of notification",
            "example": "PROJECT_INVITATION",
            "enum": [
              "PROJECT_INVITATION",
              "APPLICATION_APPROVED",
              "TASK_APPROVED",
              "MENTOR_REQUEST",
              "NEW_MESSAGE",
              "PROJECT_UPDATE"
            ]
          },
          "message": {
            "type": "string",
            "description": "Notification message content",
            "example": "You have been invited to join the AI Study Planner project"
          },
          "readStatus": {
            "type": "boolean",
            "description": "Whether the notification should be marked as read initially",
            "example": false,
            "default": false
          },
          "referenceId": {
            "type": "string",
            "description": "Reference ID (could be project ID, task ID, etc.)",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          }
        },
        "required": [
          "userId",
          "type",
          "message"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "readStatus": {
            "type": "boolean",
            "description": "Whether the notification has been read",
            "example": true
          }
        }
      },
...
    }
  }
}